==> notebooks/05_create_database_pg_schema.ipynb <==
import sys
import os

sys.path.append(os.path.abspath('..'))

print(f"Caminho '..' foi adicionado: {os.path.abspath('..')}")

from database.config.database import create_database_if_not_exists

# Garante que o banco de dados exista antes de tentar criar as tabelas
create_database_if_not_exists()

from sqlalchemy import create_engine, text
from database.config.database import get_engine

# Conectar ao banco
engine = get_engine()

# Script SQL para criar tabelas
sql_script = """
CREATE TABLE IF NOT EXISTS analise_sazonal_clima (
    "season" VARCHAR(255),
    "crop" VARCHAR(255),
    "producao_media" DECIMAL,
    "rendimento_medio" DECIMAL,
    "area_media" DECIMAL,
    "temp_media_jun_sep" DECIMAL,
    "precipitacao_media_jun_sep" DECIMAL
);

CREATE TABLE IF NOT EXISTS benchmark_regional_rendimento (
    "state_district" VARCHAR(255),
    "crop" VARCHAR(255),
    "rendimento_medio_local" DECIMAL,
    "producao_total_local" DECIMAL,
    "rendimento_medio_nacional" DECIMAL,
    "diferenca_vs_media_percent" DECIMAL
);

CREATE TABLE IF NOT EXISTS desempenho_regiao_cultura (
    "state_district" VARCHAR(255),
    "crop" VARCHAR(255),
    "area_total" DECIMAL,
    "producao_total" DECIMAL,
    "rendimento_medio_local" DECIMAL
);

CREATE TABLE IF NOT EXISTS perfil_climatico_regiao_cultura (
    "state_district" VARCHAR(255),
    "crop" VARCHAR(255),
    "rendimento_medio" DECIMAL,
    "precipitacao_anual_media" DECIMAL,
    "precipitacao_jun_sep_media" DECIMAL,
    "temp_media_janfeb" DECIMAL,
    "temp_media_marmay" DECIMAL,
    "temp_media_junsep" DECIMAL,
    "temp_media_octdec" DECIMAL
);

CREATE TABLE IF NOT EXISTS producao_anual_cultura (
    "year" INTEGER,
    "crop" VARCHAR(255),
    "area_total_plantada" DECIMAL,
    "producao_total" DECIMAL,
    "rendimento_ponderado" DECIMAL
);

CREATE TABLE IF NOT EXISTS tendencia_anual_rendimento (
    "crop" VARCHAR(255),
    "year" INTEGER,
    "rendimento_medio" DECIMAL,
    "crescimento_anual_rendimento_percent" DECIMAL
);

CREATE TABLE IF NOT EXISTS volatilidade_rendimento_regiao (
    "state_district" VARCHAR(255),
    "crop" VARCHAR(255),
    "rendimento_medio" DECIMAL,
    "rendimento_std_dev" DECIMAL,
    "contagem_registros" INTEGER,
    "coef_variacao_rendimento_percent" DECIMAL
);
"""

# Executar script
with engine.connect() as conn:
    conn.execute(text(sql_script))
    conn.commit()

print("Schema do banco criado com sucesso!")
